Authorize googlesheets4 using GCP Service Account Key

Authorizes the googlesheets4 package to access and manage Google Sheets via a Google Cloud Platform (GCP) Service Account Key file (in JSON format). See the relevant documentation for details.

Usage

auth_g_sheets_gcp(
  path_gcp_service_account_key = Sys.getenv("PATH_GCP_KEY_ZDA")
)

Arguments

path_gcp_service_account_key

Path to the GCP Service Account Key JSON file.

Value

path_gcp_service_account_key, invisibly.

Details

The recommended way to provide the filesystem path to the GCP Service Account Key file is to set the environment variable PATH_GCP_KEY_ZDA to that path, e.g. via the .Renviron file.

See also