Ballot data

A list of tibbles containing official statistical data about all FOKUS-covered ballots. The data includes i.a. the total electorate as well as participation numbers per voting channel and per voting decision.

Usage

ballots

Format

A tibble.

See also

Examples

fokus::ballots
#> $referendum
#> # A tibble: 5,513 × 18
#>    fso_id entity       ballot_date lvl   canton ballot_type proposal_nr electorate_total votes_received votes_empty votes_invalid votes_valid votes_yes votes_no
#>     <int> <chr>        <date>      <chr> <chr>  <chr>             <int>            <int>          <int>       <int>         <int>       <int>     <int>    <int>
#>  1     19 Kanton Aarg… 2018-09-23  cant… aargau referendum            1           415726         147027        1523            80      145424     36959   108465
#>  2   1901 Bezirk Aarau 2018-09-23  cant… aargau referendum            1            48918          18835         192             7       18636      5466    13170
#>  3   1902 Bezirk Baden 2018-09-23  cant… aargau referendum            1            84038          30973         357            13       30603      7386    23217
#>  4   1903 Bezirk Brem… 2018-09-23  cant… aargau referendum            1            47439          17019         155             5       16859      3771    13088
#>  5   1904 Bezirk Brugg 2018-09-23  cant… aargau referendum            1            31805          12605         137             5       12463      3310     9153
#>  6   1905 Bezirk Kulm  2018-09-23  cant… aargau referendum            1            25548           8313          90            10        8213      2095     6118
#>  7   1906 Bezirk Lauf… 2018-09-23  cant… aargau referendum            1            21195           7233          76             2        7155      1859     5296
#>  8   1907 Bezirk Lenz… 2018-09-23  cant… aargau referendum            1            39471          13713         121            12       13580      3336    10244
#>  9   1908 Bezirk Muri  2018-09-23  cant… aargau referendum            1            23791           7883          81             0        7802      1738     6064
#> 10   1909 Bezirk Rhei… 2018-09-23  cant… aargau referendum            1            29242           9370         100            21        9249      2377     6872
#> # ℹ 5,503 more rows
#> # ℹ 4 more variables: postal_votes_received <int>, postal_votes_invalid <int>, ballot_votes_received <int>, evoting_votes_received <int>
#> 
#> $election
#> $election$`2019-10-20_aargau_cantonal_majoritarian_1`
#> # A tibble: 223 × 15
#>    fso_id entity             ballot_date electorate_total votes_received votes_empty votes_invalid votes_valid `Jeanine Glarner (FDP)` `Yvonne Feri (SP)`
#>     <int> <chr>              <date>                 <int>          <int>       <int>         <int>       <int>                   <int>              <int>
#>  1     19 Kanton Aargau      2019-10-20            419164         183984        1602           359      182023                   21882              44765
#>  2   1901 Bezirk Aarau       2019-10-20             49470          24215         164            48       24003                    3020               6577
#>  3   1902 Bezirk Baden       2019-10-20             84622          38801         412            78       38311                    5301              10747
#>  4   1903 Bezirk Bremgarten  2019-10-20             47674          20559         163            71       20325                    2642               4440
#>  5   1904 Bezirk Brugg       2019-10-20             32186          15357         131            25       15201                    1705               4029
#>  6   1905 Bezirk Kulm        2019-10-20             25791          10506          55            28       10423                     894               1699
#>  7   1906 Bezirk Laufenburg  2019-10-20             21308           9026         111            16        8899                     988               2315
#>  8   1907 Bezirk Lenzburg    2019-10-20             40201          17984          97            29       17858                    1943               3931
#>  9   1908 Bezirk Muri        2019-10-20             24042           9976          99            19        9858                    1237               1874
#> 10   1909 Bezirk Rheinfelden 2019-10-20             29297          11198         148            24       11026                    1564               3008
#> # ℹ 213 more rows
#> # ℹ 5 more variables: `Severin Lüscher (Grüne)` <int>, `Pius Lischer (Neue Bundesverfassung)` <int>, `Jean-Pierre Gallati (SVP)` <int>,
#> #   `Doris Aebi (GLP)` <int>, `jemand anderes (nicht nominiert)` <int>
#> 
#> $election$`2019-10-20_aargau_federal_majoritarian_1`
#> # A tibble: 224 × 24
#>    fso_id entity          ballot_date electorate_total postal_votes_received postal_votes_invalid ballot_votes_received votes_received votes_empty votes_invalid
#>     <int> <chr>           <date>                 <int>                 <int>                <int>                 <int>          <int>       <int>         <int>
#>  1     19 Kanton Aargau   2019-10-20            429507                191532                 1872                  7488         190672         614           148
#>  2   1901 Bezirk Aarau    2019-10-20             49470                 25106                  282                   663          24769          77            10
#>  3   1902 Bezirk Baden    2019-10-20             84622                 39305                  351                  1844          39987         136            31
#>  4   1903 Bezirk Bremgar… 2019-10-20             47674                 21192                  251                   911          21062          71            16
#>  5   1904 Bezirk Brugg    2019-10-20             32186                 15692                  120                   466          15694          42             8
#>  6   1905 Bezirk Kulm     2019-10-20             25791                 10689                   74                   364          10665          19            10
#>  7   1906 Bezirk Laufenb… 2019-10-20             21308                  9312                   78                   377           9320          28             9
#>  8   1907 Bezirk Lenzburg 2019-10-20             40201                 18741                  201                   617          18406          48             9
#>  9   1908 Bezirk Muri     2019-10-20             24042                 10426                  102                   364          10309          21            12
#> 10   1909 Bezirk Rheinfe… 2019-10-20             29297                 12136                  137                   441          11547          83            21
#> # ℹ 214 more rows
#> # ℹ 14 more variables: votes_valid <int>, `Beat Flach (GLP)` <int>, `Pius Lischer (Neue Bundesverfassung)` <int>, `Thierry Burkart (FDP)` <int>,
#> #   `Cédric Wermuth (SP)` <int>, `Jean-Pierre Leutwyler (FW AG)` <int>, `Marianne Binder-Keller (CVP)` <int>, `Hansjörg Knecht (SVP)` <int>,
#> #   `Ruth Müri (Grüne)` <int>, `Maya Bally (BDP)` <int>, `Roland Frauchiger (EVP)` <int>, `jemand anderes (nicht nominiert)` <int>, isolated_votes_empty <int>,
#> #   isolated_votes_invalid <int>
#> 
#> $election$`2019-10-20_aargau_federal_proportional_1`
#> # A tibble: 224 × 19
#>    fso_id entity      ballot_date votes_total    CVP   EDU `PPAG (Piratenpartei)` `FW AG (Freie Wähler)` LOVB (Lösungs-Orient…¹ `TEAM65+` DU (Die Unabhängigen…²
#>     <int> <chr>       <date>            <int>  <int> <int>                  <int>                  <int>                  <int>     <int>                  <int>
#>  1     19 Kanton Aar… 2019-10-20      2985938 295446 29768                  10090                   6954                   4094     41127                   7651
#>  2   1901 Bezirk Aar… 2019-10-20       386897  23025  3748                   1391                    772                    490      6294                    767
#>  3   1902 Bezirk Bad… 2019-10-20       621508  75921  3160                   1962                   1711                    787      7562                   2074
#>  4   1903 Bezirk Bre… 2019-10-20       331420  40164  2210                    900                    644                    479      3850                    687
#>  5   1904 Bezirk Bru… 2019-10-20       245454  18981  2196                    895                    511                    266      3870                    786
#>  6   1905 Bezirk Kulm 2019-10-20       167095   7681  5212                    351                    440                    249      3116                    291
#>  7   1906 Bezirk Lau… 2019-10-20       146871  22930  1084                    672                    316                    200      2933                    204
#>  8   1907 Bezirk Len… 2019-10-20       286721  19273  2303                   1074                    589                    350      3371                    750
#>  9   1908 Bezirk Muri 2019-10-20       163750  30276  1044                    532                    446                    448      1330                    641
#> 10   1909 Bezirk Rhe… 2019-10-20       180556  19756  1919                   1123                    468                    246      3718                    658
#> # ℹ 214 more rows
#> # ℹ abbreviated names: ¹​`LOVB (Lösungs-Orientierte Volks-Bewegung)`, ²​`DU (Die Unabhängigen)`
#> # ℹ 8 more variables: `LS (Luzi Stamm)` <int>, `BDP oder Junge BDP` <dbl>, `EVP oder Junge EVP` <dbl>, `FDP oder Jungfreisinnige` <dbl>,
#> #   `GLP oder Junge GLP` <dbl>, `Grüne oder Junge Grüne` <dbl>, `SP oder JUSO` <dbl>, `SVP oder Junge SVP` <dbl>
#> 
#> $election$`2020-10-18_aargau_cantonal_majoritarian_1`
#> # A tibble: 222 × 22
#>    fso_id entity             ballot_date electorate_total votes_received votes_empty votes_invalid votes_valid `Alex Hürzeler (SVP)` `Stephan Attiger (FDP)`
#>     <int> <chr>              <date>                 <int>          <int>       <int>         <int>       <int>                 <int>                   <int>
#>  1     19 Kanton Aargau      2020-10-18            422612         136821         552            97      136172                 87413                   95404
#>  2   1901 Bezirk Aarau       2020-10-18             49894          19305          86            27       19192                 11772                   13107
#>  3   1902 Bezirk Baden       2020-10-18             84965          27795          99            13       27683                 16259                   19651
#>  4   1903 Bezirk Bremgarten  2020-10-18             48278          14866          66             9       14791                  9468                   10315
#>  5   1904 Bezirk Brugg       2020-10-18             32381          11900          39             8       11853                  7528                    8352
#>  6   1905 Bezirk Kulm        2020-10-18             26012           7758          39             4        7715                  5798                    5732
#>  7   1906 Bezirk Laufenburg  2020-10-18             21563           7418          30             1        7387                  5271                    5178
#>  8   1907 Bezirk Lenzburg    2020-10-18             40827          13213          48            11       13154                  8681                    9352
#>  9   1908 Bezirk Muri        2020-10-18             24388           7053          24             6        7023                  4796                    5097
#> 10   1909 Bezirk Rheinfelden 2020-10-18             29327           7641          39            13        7589                  4555                    4696
#> # ℹ 212 more rows
#> # ℹ 12 more variables: `Markus Dieth (CVP)` <int>, `Jean-Pierre Gallati (SVP)` <int>, `Cybel Dickson (JUSO)` <int>, `Dieter Egli (SP)` <int>,
#> #   `Christiane Guyer (Grüne)` <int>, `Patricia Hegglin (JUSO)` <int>, `Pius Lischer (Für Freiheit und Gesundheit)` <int>, `Theres Schöni (LOVB)` <int>,
#> #   `Zoe Sutter (JUSO)` <int>, `Stephan Zurfluh (i54.ch)` <int>, `jemand anderes (nicht nominiert)` <int>, slots_empty <int>
#> 
#> $election$`2020-10-18_aargau_cantonal_proportional_1`
#> # A tibble: 12 × 15
#>    entity          electorate_total turnout votes_total    SVP `SP / JUSO` `FDP / Jungfreisinnige`    CVP `Grüne / Junge Grüne`    GLP    EVP    EDU
#>    <chr>                      <int>   <dbl>       <dbl>  <dbl>       <dbl>                   <dbl>  <dbl>                 <dbl>  <dbl>  <dbl>  <dbl>
#>  1 Aarau                      49894    39.3      19048.  4914.       3804.                   3437   1225.                 2230.  1614. 1080.   316. 
#>  2 Baden                      84965    33.1      26936.  6508.       4549.                   3759.  4476.                 3121.  3255. 1021.   131. 
#>  3 Bremgarten                 48278    31.4      14543.  4907.       2229.                   2086.  2205.                 1170.  1581.  230.    89.1
#>  4 Brugg                      32381    37.3      11746.  3388        2099.                   2094.  1100.                 1327.   799.  764    175. 
#>  5 Kulm                       26012    30.7       7781.  3074.        921.                   1388.   196.                  766.   276.  557.   585. 
#>  6 Laufenburg                 21563    35.3       7416.  2417        1475                     745.  1419.                  597.   457.  135.   114. 
#>  7 Lenzburg                   40827    32.8      12908.  4388.       1997.                   2105.  1140.                 1211.  1337.  590.    80.9
#>  8 Muri                       24388    29.8       7126.  2352.        795.                    644.  1672.                  467.   990.   94.7   53.4
#>  9 Rheinfelden                29327    27.2       7670.  2150.       1489.                   1144.  1034.                  799.   842.  114.    98  
#> 10 Zofingen                   44800    30.1      12963.  4416.       2140.                   1758.   868.                 1313.   988. 1012.   467. 
#> 11 Zurzach                    20177    34.8       6824.  2385.        837.                    690.  1943.                  507.   321.   73.6   51.9
#> 12 Total im Kanton           422612    33.0     134962. 40900.      22335.                  19850. 17278.                13506. 12459. 5671.  2162. 
#> # ℹ 3 more variables: `LOVB (Lösungs-Orientierte Volks-Bewegung)` <dbl>, `FFF (Frecher Frischer Fischer)` <dbl>, `PPAG (Piratenpartei)` <dbl>
#> 
#>