Get referendum proposal combinations for which arguments have been queried

Returns a list with a metadata item per ballot date, political level, canton, proposal number and optionally proposal argument side (pro/contra) and number.

Usage

combos_proposal_arguments(
  ballot_dates = all_ballot_dates,
  lvls = all_lvls,
  cantons = all_cantons,
  incl_side = TRUE,
  incl_argument_nr = incl_side
)

Arguments

ballot_dates

FOKUS-covered ballot date(s). One or more of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

"2023-06-18"

lvls

Political level(s). One or more of "cantonal" and "federal".

cantons

FOKUS-covered canton name(s). One or more of

"aargau"

incl_side

Whether or not to include argument sides (pro/contra) in the resulting list. Setting this to FALSE potentially results in fewer combinations.

incl_argument_nr

Whether or not to include argument numbers in the resulting list. Setting this to FALSE potentially results in fewer combinations. Setting this to TRUE implies incl_side = TRUE.

Value

A list with an element per political-level-proposal-number and optionally -argument-side and -argument-number combination.

See also

Examples

fokus::combos_proposal_arguments(ballot_dates = "2023-06-18",
                                 cantons = "aargau")
#> [[1]]
#> [[1]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[1]]$lvl
#> [1] "cantonal"
#> 
#> [[1]]$canton
#> [1] "aargau"
#> 
#> [[1]]$proposal_nr
#> [1] 1
#> 
#> [[1]]$side
#> [1] "pro"
#> 
#> [[1]]$argument_nr
#> [1] 1
#> 
#> 
#> [[2]]
#> [[2]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[2]]$lvl
#> [1] "cantonal"
#> 
#> [[2]]$canton
#> [1] "aargau"
#> 
#> [[2]]$proposal_nr
#> [1] 1
#> 
#> [[2]]$side
#> [1] "contra"
#> 
#> [[2]]$argument_nr
#> [1] 1
#> 
#> 
#> [[3]]
#> [[3]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[3]]$lvl
#> [1] "cantonal"
#> 
#> [[3]]$canton
#> [1] "aargau"
#> 
#> [[3]]$proposal_nr
#> [1] 1
#> 
#> [[3]]$side
#> [1] "pro"
#> 
#> [[3]]$argument_nr
#> [1] 2
#> 
#> 
#> [[4]]
#> [[4]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[4]]$lvl
#> [1] "cantonal"
#> 
#> [[4]]$canton
#> [1] "aargau"
#> 
#> [[4]]$proposal_nr
#> [1] 1
#> 
#> [[4]]$side
#> [1] "contra"
#> 
#> [[4]]$argument_nr
#> [1] 2
#> 
#> 
#> [[5]]
#> [[5]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[5]]$lvl
#> [1] "cantonal"
#> 
#> [[5]]$canton
#> [1] "aargau"
#> 
#> [[5]]$proposal_nr
#> [1] 3
#> 
#> [[5]]$side
#> [1] "pro"
#> 
#> [[5]]$argument_nr
#> [1] 1
#> 
#> 
#> [[6]]
#> [[6]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[6]]$lvl
#> [1] "cantonal"
#> 
#> [[6]]$canton
#> [1] "aargau"
#> 
#> [[6]]$proposal_nr
#> [1] 3
#> 
#> [[6]]$side
#> [1] "contra"
#> 
#> [[6]]$argument_nr
#> [1] 1
#> 
#> 
#> [[7]]
#> [[7]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[7]]$lvl
#> [1] "cantonal"
#> 
#> [[7]]$canton
#> [1] "aargau"
#> 
#> [[7]]$proposal_nr
#> [1] 3
#> 
#> [[7]]$side
#> [1] "pro"
#> 
#> [[7]]$argument_nr
#> [1] 2
#> 
#> 
#> [[8]]
#> [[8]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[8]]$lvl
#> [1] "cantonal"
#> 
#> [[8]]$canton
#> [1] "aargau"
#> 
#> [[8]]$proposal_nr
#> [1] 3
#> 
#> [[8]]$side
#> [1] "contra"
#> 
#> [[8]]$argument_nr
#> [1] 2
#> 
#> 
#> [[9]]
#> [[9]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[9]]$lvl
#> [1] "cantonal"
#> 
#> [[9]]$canton
#> [1] "aargau"
#> 
#> [[9]]$proposal_nr
#> [1] 3
#> 
#> [[9]]$side
#> [1] "pro"
#> 
#> [[9]]$argument_nr
#> [1] 3
#> 
#> 
#> [[10]]
#> [[10]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[10]]$lvl
#> [1] "federal"
#> 
#> [[10]]$canton
#> [1] "aargau"
#> 
#> [[10]]$proposal_nr
#> [1] 2
#> 
#> [[10]]$side
#> [1] "pro"
#> 
#> [[10]]$argument_nr
#> [1] 1
#> 
#> 
#> [[11]]
#> [[11]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[11]]$lvl
#> [1] "federal"
#> 
#> [[11]]$canton
#> [1] "aargau"
#> 
#> [[11]]$proposal_nr
#> [1] 2
#> 
#> [[11]]$side
#> [1] "contra"
#> 
#> [[11]]$argument_nr
#> [1] 1
#> 
#> 
#> [[12]]
#> [[12]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[12]]$lvl
#> [1] "federal"
#> 
#> [[12]]$canton
#> [1] "aargau"
#> 
#> [[12]]$proposal_nr
#> [1] 2
#> 
#> [[12]]$side
#> [1] "pro"
#> 
#> [[12]]$argument_nr
#> [1] 2
#> 
#> 
#> [[13]]
#> [[13]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[13]]$lvl
#> [1] "federal"
#> 
#> [[13]]$canton
#> [1] "aargau"
#> 
#> [[13]]$proposal_nr
#> [1] 2
#> 
#> [[13]]$side
#> [1] "contra"
#> 
#> [[13]]$argument_nr
#> [1] 2
#> 
#> 
# without argument numbers
fokus::combos_proposal_arguments(ballot_dates = "2023-06-18",
                                 cantons = "aargau",
                                 incl_argument_nr = FALSE)
#> [[1]]
#> [[1]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[1]]$lvl
#> [1] "cantonal"
#> 
#> [[1]]$canton
#> [1] "aargau"
#> 
#> [[1]]$proposal_nr
#> [1] 1
#> 
#> [[1]]$side
#> [1] "pro"
#> 
#> 
#> [[2]]
#> [[2]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[2]]$lvl
#> [1] "cantonal"
#> 
#> [[2]]$canton
#> [1] "aargau"
#> 
#> [[2]]$proposal_nr
#> [1] 1
#> 
#> [[2]]$side
#> [1] "contra"
#> 
#> 
#> [[3]]
#> [[3]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[3]]$lvl
#> [1] "cantonal"
#> 
#> [[3]]$canton
#> [1] "aargau"
#> 
#> [[3]]$proposal_nr
#> [1] 3
#> 
#> [[3]]$side
#> [1] "pro"
#> 
#> 
#> [[4]]
#> [[4]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[4]]$lvl
#> [1] "cantonal"
#> 
#> [[4]]$canton
#> [1] "aargau"
#> 
#> [[4]]$proposal_nr
#> [1] 3
#> 
#> [[4]]$side
#> [1] "contra"
#> 
#> 
#> [[5]]
#> [[5]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[5]]$lvl
#> [1] "federal"
#> 
#> [[5]]$canton
#> [1] "aargau"
#> 
#> [[5]]$proposal_nr
#> [1] 2
#> 
#> [[5]]$side
#> [1] "pro"
#> 
#> 
#> [[6]]
#> [[6]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[6]]$lvl
#> [1] "federal"
#> 
#> [[6]]$canton
#> [1] "aargau"
#> 
#> [[6]]$proposal_nr
#> [1] 2
#> 
#> [[6]]$side
#> [1] "contra"
#> 
#> 
# without argument sides and numbers
fokus::combos_proposal_arguments(ballot_dates = "2023-06-18",
                                 cantons = "aargau",
                                 incl_side = FALSE,
                                 incl_argument_nr = FALSE)
#> [[1]]
#> [[1]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[1]]$lvl
#> [1] "cantonal"
#> 
#> [[1]]$canton
#> [1] "aargau"
#> 
#> [[1]]$proposal_nr
#> [1] 1
#> 
#> 
#> [[2]]
#> [[2]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[2]]$lvl
#> [1] "cantonal"
#> 
#> [[2]]$canton
#> [1] "aargau"
#> 
#> [[2]]$proposal_nr
#> [1] 3
#> 
#> 
#> [[3]]
#> [[3]]$ballot_date
#> [1] "2023-06-18"
#> 
#> [[3]]$lvl
#> [1] "federal"
#> 
#> [[3]]$canton
#> [1] "aargau"
#> 
#> [[3]]$proposal_nr
#> [1] 2
#> 
#>