Export all questionnaires

Exports all questionnaires, softly by default (i.e. without an XLSX version, without deploying as a static site and without uploading to Google Drive).

export_q_all(
  verbose = FALSE,
  incl_csv = TRUE,
  incl_html = TRUE,
  incl_xlsx = FALSE,
  deploy = FALSE,
  local_deploy_path = getOption("fokus.q.local_deploy_path"),
  upload_to_g_drive = FALSE,
  g_drive_folder = "fokus/aargau/Umfragen/Dateien für Umfrageinstitut/Fragebogen/"
)

Arguments

verbose
Whether or not to print progress information during questionnaire generation.

incl_csv
Whether or not to also generate and export a CSV version of the questionnaire.

incl_html
Whether or not to also generate and export an HTML version of the questionnaire.

incl_xlsx
Whether or not to also generate and export an XLSX version of the questionnaire.

deploy
Whether or not to deploy the generated files as a static site to the Git repository specified under local_deploy_path.

local_deploy_path
Local filesystem path to the Git repository of the static site to deploy the generated files files to. Ignored if deploy = FALSE.

upload_to_g_drive
Whether or not to upload the generated files to the Google Drive folder g_drive_folder.

g_drive_folder
Google Drive folder to deploy the generated files to. Ignored if upload_to_g_drive = FALSE.

Details

Useful to efficiently test and inspect latest changes in generated questionnaire files.

See also