Export QR codes with personalized survey URL

Exports a ZIP file, that contains a QR code in SVG and in EPS format for each survey participant storing the personalized survey URL, to the private FOKUS repository.

Usage

export_qr_codes(
  ballot_date = pal::pkg_config_val("ballot_date"),
  canton = cantons(ballot_date),
  upload_to_g_drive = TRUE,
  g_drive_folder =
    glue::glue("fokus/{canton}/Umfragen/Dateien für Umfrageinstitut/QR-Codes/"),
  auth_token = pal::pkg_config_val("token_repo_private"),
  quiet = FALSE,
  verbose = FALSE
)

Arguments

ballot_date

FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

"2023-06-18"

"2024-10-20"

canton

FOKUS-covered canton name. One of

"aargau"

upload_to_g_drive

Whether or not to upload the generated files to the Google Drive folder g_drive_folder.

g_drive_folder

Google Drive folder to upload the generated files to. Ignored if upload_to_g_drive = FALSE.

auth_token

Personal access token of a gitlab.com account with access to the private FOKUS repository.

quiet

Whether or not to suppress printing status output from internal processing.

verbose

Whether or not to print detailed status output from Google Drive file upload.

Value

A tibble containing metadata about the contents of the created ZIP archive, invisibly.

See also