fokus: Provides an API around the FOKUS Post-voting Surveys

fokus allows to (pre-)process and clean the raw data, analyze and visualize the cleaned data, create the questionnaires and provides other utility functions around the FOKUS post-voting surveys by the Centre for Democracy Studies Aarau (ZDA) at the University of Zurich, Switzerland.

Package configuration

Some of fokus's functionality is controlled via package-specific global configuration which can either be set via R options or environment variables (the former take precedence). This configuration includes:

Description	R option	Environment variable	Default value

FOKUS-covered ballot date to fall back to in various functions of this package when it isn't explicitly specified. Basically a means to globally define the ballot date to be processed. One of "2018-09-23", "2018-11-25", "2019-10-20", "2020-09-27", "2020-10-18", "2021-11-28" or "2023-06-18".	fokus.ballot_date	R_FOKUS_BALLOT_DATE	2018-09-23

FOKUS-covered canton name to fall back to in various functions of this package when it isn't explicitly specified. Basically a means to globally define the canton to be processed. One of "aargau".	fokus.canton	R_FOKUS_CANTON	"aargau"

Maximal timespan to preserve the package's pkgpins cache. Cache entries older than this will be deleted upon package loading.	fokus.global_max_cache_age	R_FOKUS_GLOBAL_MAX_CACHE_AGE	"1 day"

Personal access token of a gitlab.com account with access to the private FOKUS repository.	fokus.token_repo_private	R_FOKUS_TOKEN_REPO_PRIVATE	Sys.getenv("GITLAB_COM_TOKEN")

See also

Author

Maintainer: Salim Brüggemann salim-b@pm.me (ORCID)

Other contributors:

Centre for Democracy Studies Aarau (ZDA) info@zdaarau.ch [copyright holder]