Determine whether ballot includes type

Determines whether or not the FOKUS survey for the specified canton at the specified ballot date on the specified political level(s) covered the specified ballot types.

Usage

has_ballot_type(
  ballot_date = pal::pkg_config_val("ballot_date"),
  lvls = all_lvls,
  canton = cantons(ballot_date),
  ballot_type = ballot_types(ballot_date = ballot_date, lvls = lvls, canton = canton),
  prcds = all_prcds
)

Arguments

ballot_date

FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

"2023-06-18"

"2024-10-20"

lvls

Political level(s). One or more of "cantonal" and "federal".

canton

FOKUS-covered canton name. One of

"aargau"

Only relevant if lvls includes "cantonal" or ballot_type = "election".

ballot_type

Ballot type. One of "referendum" or "election".

prcds

Election procedure(s). One or more of "proportional" and "majoritarian".

Only relevant if ballot_type = "election".

Value

A logical vector of the same length as lvl (× prcds), named after lvls(.prcds).ballot_type.

See also

Examples

fokus::has_ballot_type(ballot_date = "2018-09-23",
                       canton = "aargau")
#> cantonal.referendum  federal.referendum 
#>                TRUE                TRUE 
                        
fokus::has_ballot_type(ballot_date = "2018-09-23",
                       canton = "aargau",
                       ballot_type = "election")
#> cantonal.proportional.election cantonal.majoritarian.election  federal.proportional.election  federal.majoritarian.election 
#>                          FALSE                          FALSE                          FALSE                          FALSE 
                        
fokus::has_ballot_type(ballot_date = "2019-10-20",
                       canton = "aargau",
                       ballot_type = "election",
                       prcds = "proportional")
#> cantonal.proportional.election  federal.proportional.election 
#>                          FALSE                           TRUE