Determine whether ballot type includes an election

Determines whether or not the ballot for the specified canton at the specified date on the specified political level(s) includes an election of the specified election procedure(s).

has_election(
  ballot_date = ballot_dates,
  lvls = c("cantonal", "federal"),
  canton = cantons,
  prcds = c("proportional", "majoritarian")
)

Arguments

ballot_date
A valid FOKUS-covered cantonal ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

lvls
Political level(s). One or more of

"cantonal"

"federal"

canton
A valid FOKUS canton. One of

"aargau"

prcds
Election procedure(s). One or more of

"proportional"

"majoritarian"

Value

A logical scalar.

Details

Note that if more than one lvls/prcds are provided (the default), they are interpreted according to the rules of logical disjunction, i.e. TRUE is returned if any lvls/prcds are included in the specified ballot.

For logical conjunction, use e.g. has_election(lvls = "cantonal") && has_election(lvls = "federal").

See also

Examples

fokus::has_election(ballot_date = "2018-09-23",
                    canton = "aargau")

#> [1] FALSE