Determine whether ballot includes elections

Determines whether or not the FOKUS survey for the specified canton at the specified ballot date on the specified political levels covered the specified election numbers of the specified procedures.

has_election_nrs(
  ballot_date = all_ballot_dates,
  lvls = all_lvls,
  canton = cantons(ballot_date),
  prcds = all_prcds,
  election_nrs = NULL
)

Arguments

ballot_date
FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

lvls
Political level(s). One or more of

"cantonal"

"federal"

canton
Canton name. One of

"aargau"

prcds
Election procedure(s). One or more of

"proportional"

"majoritarian"

election_nrs
Election number(s) to test for. An integerish vector or NULL. If NULL, falls back to 1:n_elections() present matching the specified parameters.

Value

Either

an unnamed logical scalar if election_nrs is NULL and there was no FOKUS-covered election of any prcds at ballot_date on any lvls, or

a logical vector of the same length as lvls × prcds × election_nrs, named after lvls.prcds.election_nrs.

See also

Examples

fokus::has_election_nrs(ballot_date = "2019-10-20",
                        canton = "aargau")
#> cantonal.majoritarian.1  federal.proportional.1  federal.majoritarian.1 
#>                    TRUE                    TRUE                    TRUE 

fokus::has_election_nrs(ballot_date = "2019-10-20",
                        lvls = "federal",
                        canton = "aargau",
                        prcds = "proportional")
#> federal.proportional.1 
#>                   TRUE 
                        
fokus::has_election_nrs(ballot_date = "2019-10-20",
                        canton = "aargau",
                        election_nrs = 1:2)
#> cantonal.proportional.1 cantonal.proportional.2 cantonal.majoritarian.1 
#>                   FALSE                   FALSE                    TRUE 
#> cantonal.majoritarian.2  federal.proportional.1  federal.proportional.2 
#>                   FALSE                    TRUE                   FALSE 
#>  federal.majoritarian.1  federal.majoritarian.2 
#>                    TRUE                   FALSE 

fokus::has_election_nrs(ballot_date = "2018-09-23",
                        canton = "aargau")
#> [1] FALSE

fokus::has_election_nrs(ballot_date = "2020-10-18",
                        lvls = "federal",
                        canton = "aargau")
#> [1] FALSE