Determine whether ballot includes a referendum

Determines whether or not the FOKUS survey for the specified canton at the specified ballot date on the specified political level(s) covered a referendum.

Usage

has_referendum(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvls = all_lvls,
  canton = cantons(ballot_date)
)

Arguments

ballot_date

FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

"2023-06-18"

lvls

Political level(s). One or more of "cantonal" and "federal".

canton

FOKUS-covered canton name. One of

"aargau"

Only relevant if lvls includes "cantonal".

Value

A logical vector of the same length as and named after lvls.

See also

Examples

fokus::has_referendum(ballot_date = "2018-09-23",
                      canton = "aargau")
#> cantonal  federal 
#>     TRUE     TRUE 

fokus::has_referendum(ballot_date = "2018-09-23",
                      lvls = "federal",
                      canton = "aargau")
#> federal 
#>    TRUE