Determine whether variable is skill question

Determines for each variable whether or not it is a skill question.

Usage

is_skill_question_var(var_names)

Arguments

var_names

A character vector of FOKUS variable names.

Value

A logical vector of the same length as var_names.

Details

Note that the determination is performed by simply parsing var_names.

See also