Get proposal's arguments

Returns text, side and number of all arguments on the specified proposal.

Usage

proposal_arguments(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = lvls(ballot_date, canton, ballot_type = "referendum"),
  canton = cantons(ballot_date),
  proposal_nr = 1L
)

Arguments

ballot_date

FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

"2023-06-18"

lvl

Political level. One of "cantonal" or "federal".

canton

FOKUS-covered canton name. One of

"aargau"

Only relevant if lvl = "cantonal".

proposal_nr

Proposal number. A positive integerish scalar.

Value

A tibble.

See also

Examples

fokus::proposal_arguments(ballot_date = "2018-09-23",
                          lvl = "cantonal",
                          canton = "aargau",
                          proposal_nr = 1)
#> # A tibble: 6 × 5
#>   de.long                                                                                                                           de.short en.long    nr side 
#>   <chr>                                                                                                                             <chr>    <chr>   <int> <chr>
#> 1 Eine Erhöhung der Vermögenssteuer für Reiche ist ein geeignetes Mittel, um den Kantonshaushalt wieder ins Lot zu bringen.         Vermöge… An inc…     1 pro  
#> 2 Eine Erhöhung der Vermögenssteuer für Reiche bringt nichts, denn Wohlhabende finden immer ein Schlupfloch, um sich ihrer Steuerp… Wohlhab… An inc…     1 cont…
#> 3 Die Reichen wurden in der Vergangenheit nicht ausreichend in die finanzielle Verantwortung genommen. Sie sollten sich stärker al… Reiche … In the…     2 pro  
#> 4 Der Kanton Aargau braucht höhere Steuereinnahmen, um all seine öffentlichen Aufgaben ohne Leistungsabbau erfüllen zu können.      Höhere … The ca…     3 pro  
#> 5 Wird der Vermögenssteuersatz für Reiche erhöht, ziehen diese aus dem Kanton weg. Dies führt zu Steuerausfällen statt höheren Ein… Reiche … If the…     2 cont…
#> 6 Die Steuern im Kanton Aargau sollten generell möglichst tief sein.                                                                Steuern… Taxes …     3 cont…