Get proposal's main motives

Returns text and code number of all main motives on the specified proposal.

proposal_main_motives(
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  canton = cantons,
  proposal_nr = 1L,
  type = c("yes", "no")
)

Arguments

ballot_date
A valid FOKUS-covered cantonal ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

lvl
Political level. One of

"cantonal"

"federal"

canton
A valid FOKUS canton. One of

"aargau"

proposal_nr
Proposal number. A positive integer scalar.

type
Main motive type. One of

"yes"

"no"

Value

A tibble.

See also

Examples

fokus::proposal_main_motives(ballot_date = "2018-09-23",
                             lvl = "cantonal",
                             canton = "aargau",
                             proposal_nr = 1,
                             type = "no")

#> # A tibble: 7 × 3
#>    code de                                             en                       
#>   <int> <chr>                                          <chr>                    
#> 1     1 einer bestimmten Abstimmungsempfehlung folgen  to follow a specific vot…
#> 2     2 kein Risiko eingehen                           to avoid any risks       
#> 3     3 dafür sorgen, dass alles so bleibt, wie es ist to keep things the way t…
#> 4     4 gute Steuerzahler nicht vertreiben             not to scare away good t…
#> 5     5 den Wohlstand des Aargaus sichern              to secure the prosperity…
#> 6     6 die Steuerattraktivität des Aargaus erhalten   to maintain the tax attr…
#> 7     7 nicht einer JUSO-Initiative zustimmen          didn't want to support a…