Get a political level's raw supplemental date-specific FOKUS questionnaire data

Returns a structured list of the raw supplemental date-specific FOKUS questionnaire data for the specified ballot date and political level.

q_suppl_lvl(ballot_date = ballot_dates, lvl = c("cantonal", "federal"))

Arguments

ballot_date
A valid FOKUS-covered cantonal ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

lvl
The political level. One of

"cantonal"

"federal"

Value

A strict list.

See also

Examples

fokus::q_suppl_lvl(ballot_date = "2018-09-23",
                   lvl = "cantonal") |>
  names()

#> [1] "aargau"