Get a single skill question's raw supplemental questionnaire data

Returns a structured list of a single skill question's data from the raw supplemental date-specific FOKUS questionnaire data.

raw_q_suppl_skill_question(
  ballot_date = all_ballot_dates,
  lvl = all_lvls,
  canton = all_cantons,
  proposal_nr = NULL,
  skill_question_nr = 1L
)

Arguments

ballot_date
FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

lvl
Political level. One of

"cantonal"

"federal"

canton
Canton name. One of

"aargau"

proposal_nr
Proposal number. A positive integerish scalar or NULL. If NULL, it is considered to be a non-proposal-specific skill question (the case at elections).

skill_question_nr
Skill question number. A positive integerish scalar.

Value

A strict list.

See also

Examples

fokus:::raw_q_suppl_skill_question(ballot_date = "2018-09-23",
                                   lvl = "cantonal",
                                   canton = "aargau",
                                   proposal_nr = 1,
                                   skill_question_nr = 1)
#> $de
#> [1] "Die Zahl der AargauerInnen, die Vermögenssteuer zahlen müssen (d. h. ein Vermögen besitzen, das über dem Freibetrag liegt), würde sich durch die Initiative …"
#> 
#> $en
#> [1] "With the initiative, the number of people in Aargau who have to pay wealth tax (i.e. own more than the tax-free amount) would be …"
#> 
#> $response_option
#> $response_option[[1]]
#> $response_option[[1]]$de
#> [1] "verdoppeln"
#> 
#> $response_option[[1]]$en
#> [1] "twice the size"
#> 
#> $response_option[[1]]$is_correct
#> [1] FALSE
#> 
#> 
#> $response_option[[2]]
#> $response_option[[2]]$de
#> [1] "nicht verändern"
#> 
#> $response_option[[2]]$en
#> [1] "the same"
#> 
#> $response_option[[2]]$is_correct
#> [1] TRUE
#> 
#> 
#> $response_option[[3]]
#> $response_option[[3]]$de
#> [1] "halbieren"
#> 
#> $response_option[[3]]$en
#> [1] "half the size"
#> 
#> $response_option[[3]]$is_correct
#> [1] FALSE
#> 
#> 
#>