Get a proposal's raw main motive supplemental questionnaire data

Returns a structured list of main motive data from the proposal-specific raw supplemental date-specific FOKUS questionnaire data.

Usage

raw_qstnr_suppl_main_motives(
  ballot_date = pal::pkg_config_val("ballot_date"),
  lvl = all_lvls,
  canton = pal::pkg_config_val("canton"),
  proposal_nr = 1L
)

Arguments

ballot_date

FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

"2023-06-18"

"2024-10-20"

lvl

Political level. One of "cantonal" or "federal".

canton

FOKUS-covered canton name. One of

"aargau"

proposal_nr

Proposal number. A positive integerish scalar.

Value

A strict list.

See also

Examples

fokus:::raw_qstnr_suppl_main_motives(ballot_date = "2018-11-25",
                                     lvl = "cantonal",
                                     canton = "aargau",
                                     proposal_nr = 2)
#> $no
#> $no[[1]]
#> $no[[1]]$code
#> [1] 1
#> 
#> $no[[1]]$de
#> [1] "wollte in erster Linie einer bestimmten Abstimmungsempfehlung folgen."
#> 
#> $no[[1]]$en
#> [1] "to follow a specific voting recommendation"
#> 
#> 
#> $no[[2]]
#> $no[[2]]$code
#> [1] 2
#> 
#> $no[[2]]$de
#> [1] "bin der Meinung, dass bei kantonalen Wahlen nur teilnehmen dürfen soll, wer auch im Kanton wohnhaft ist"
#> 
#> $no[[2]]$en
#> [1] "to believe that only those who reside in the canton should be allowed to take part in cantonal elections"
#> 
#> 
#> $no[[3]]
#> $no[[3]]$code
#> [1] 3
#> 
#> $no[[3]]$de
#> [1] "wollte dafür sorgen, dass alles so bleibt, wie es ist"
#> 
#> $no[[3]]$en
#> [1] "to keep things the way they are"
#> 
#> 
#> $no[[4]]
#> $no[[4]]$code
#> [1] 4
#> 
#> $no[[4]]$de
#> [1] "bin generell gegen ein Wahlrecht für Auslandschweizer/innen"
#> 
#> $no[[4]]$en
#> [1] "to be opposed in general to the right of Swiss abroad to vote"
#> 
#> 
#> $no[[5]]
#> $no[[5]]$code
#> [1] 5
#> 
#> $no[[5]]$de
#> [1] "finde, Auslandschweizer/innen sind zu weit weg vom Aargau. Sie sind mit den Gegebenheiten des Kantons nicht vertraut"
#> 
#> $no[[5]]$en
#> [1] "to believe that the Swiss abroad are too far away from Aargau and that they are not familiar with the situation in the canton"
#> 
#> 
#> 
#> $yes
#> $yes[[1]]
#> $yes[[1]]$code
#> [1] 1
#> 
#> $yes[[1]]$de
#> [1] "wollte in erster Linie einer bestimmten Abstimmungsempfehlung folgen."
#> 
#> $yes[[1]]$en
#> [1] "to follow a specific voting recommendation"
#> 
#> 
#> $yes[[2]]
#> $yes[[2]]$code
#> [1] 2
#> 
#> $yes[[2]]$de
#> [1] "wollte sicherstellen, dass Auslandschweizer/innen die gleichen Wahlrechte haben wie Inlandschweizer/innen"
#> 
#> $yes[[2]]$en
#> [1] "to ensure that Swiss abroad have the same voting rights as Swiss nationals"
#> 
#> 
#> $yes[[3]]
#> $yes[[3]]$code
#> [1] 3
#> 
#> $yes[[3]]$de
#> [1] "bin der Meinung, dass wer bei Nationalratswahlen teilnehmen kann, auch bei Ständeratswahlen soll teilnehmen dürfen"
#> 
#> $yes[[3]]$en
#> [1] "to be of the opinion that anyone who can participate in elections to the National Council should also be allowed to participate in elections to the Council of States"
#> 
#> 
#> $yes[[4]]
#> $yes[[4]]$code
#> [1] 4
#> 
#> $yes[[4]]$de
#> [1] "finde, das ist eine Selbstverständlichkeit"
#> 
#> $yes[[4]]$en
#> [1] "to think that's a matter of course"
#> 
#> 
#>