Read in easyvote municipality data

Reads in the latest dataset of easyvote municipality information provided to us not earlier than 90 days before and up until 20 days after the ballot_date.

read_easyvote_municipalities(ballot_date, canton)

Arguments

ballot_date
A valid FOKUS-covered cantonal ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

canton
A valid FOKUS canton. One of

"aargau"

Value

A tibble.

Details

If both columns min_age and max_age are NA in the data returned, it means that

the municipality did not provide easyvote with specific information on the target age range, and

the municipality has subscribed to parcel mailing (instead of direct delivery to households) and delivers the brochures itself -- very likely to young adults between 18--25 years.

Examples

# private FOKUS directory needs to be accessible for this function to work
try(
  fokus:::read_easyvote_municipalities(ballot_date = "2020-09-27",
                                       canton = "aargau")
)
#> Error in (function (cond)  : 
#>   error in evaluating the argument 'x' in selecting a method for function 'as_date': The option fokus.path_private is unset, thus we fall back to
#> /home/salim/Arbeit/ZDA/Git/zdaarau/rpkgs/fokus/docs/dev/reference. This doesn't
#> seem to be a valid FOKUS working directory. Please correct this in order for
#> this package to work properly.