Read in generated survey data

Reads in the FOKUS survey data generated by export_survey_data() for the specified ballot date in the specified canton and the specified language.

Usage

read_survey_data(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  canton = cantons(ballot_date),
  lang = all_langs
)

Arguments

ballot_date

FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

"2023-06-18"

canton

FOKUS-covered canton name. One of

"aargau"

lang

Language. One of "de" or "en".

Value

A tibble.

See also

Examples

# GitLab PAT with access to the private FOKUS repository is required for this function to work
try(
  fokus::read_survey_data(ballot_date = "2023-06-18",
                          canton = "aargau",
                          lang = "en")
)
#> # A tibble: 2,309 × 354
#>    difficulty_federal_proposal_2 explicit_consent                year_of_birth sex   cantonal_attachment political_interest_c…¹ political_interest participation
#>    <ord>                         <fct>                                   <int> <fct> <ord>               <ord>                  <ord>              <fct>        
#>  1 NA                            Yes, I am participating in thi…          1979 male  1                   rather uninterested    very interested    participated 
#>  2 NA                            Yes, I am participating in thi…          1950 male  3                   rather interested      rather interested  participated 
#>  3 NA                            NA                                       1932 male  4 (strongly attach… very interested        rather interested  participated 
#>  4 NA                            Yes, I am participating in thi…          1938 male  4 (strongly attach… very interested        very interested    participated 
#>  5 NA                            Yes, I am participating in thi…          1947 fema… 3                   rather interested      very interested    participated 
#>  6 NA                            Yes, I am participating in thi…          1952 fema… 3                   rather uninterested    rather uninterest… not particip…
#>  7 NA                            Yes, I am participating in thi…          1956 fema… 4 (strongly attach… very interested        very interested    participated 
#>  8 NA                            Yes, I am participating in thi…          1971 fema… 3                   rather interested      rather interested  participated 
#>  9 NA                            Yes, I am participating in thi…          1965 fema… 4 (strongly attach… rather interested      very interested    participated 
#> 10 NA                            Yes, I am participating in thi…          1994 male  4 (strongly attach… rather uninterested    rather interested  not particip…
#> # ℹ 2,299 more rows
#> # ℹ abbreviated name: ¹​political_interest_cantonal
#> # ℹ 346 more variables: reason_non_participation_indecision <fct>, reason_non_participation_cantonal_indecision <fct>,
#> #   reason_non_participation_disinterest <fct>, reason_non_participation_cantonal_disinterest <fct>, reason_non_participation_complexity <fct>,
#> #   reason_non_participation_cantonal_complexity <fct>, reason_non_participation_hindrance <fct>, reason_non_participation_ineffectiveness <fct>,
#> #   reason_non_participation_cantonal_ineffectiveness <fct>, reason_non_participation_insignificance <fct>,
#> #   reason_non_participation_cantonal_insignificance <fct>, reason_non_participation_forgetfulness <fct>, …