Read in voting register identifiers

Reads in the voting register identifiers of the population that was invited to participate in the FOKUS survey at the specified ballot date in the specified canton.

Usage

read_voting_register_ids(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  canton = cantons(ballot_date)
)

Arguments

ballot_date

FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

"2023-06-18"

canton

FOKUS-covered canton name. One of

"aargau"

Value

A tibble with the column id_voting_register.

Details

Note that this data is not available for all FOKUS surveys.

See also

Examples

# GitLab PAT with access to the private FOKUS repository is required for this function to work
try(
  fokus::read_voting_register_ids(ballot_date = "2019-10-20",
                                  canton = "aargau")
)
#> # A tibble: 10,000 × 1
#>    id_voting_register
#>                 <int>
#>  1             100340
#>  2             100374
#>  3             100416
#>  4             100478
#>  5             100670
#>  6             100691
#>  7             100727
#>  8             100762
#>  9             100804
#> 10             100878
#> # ℹ 9,990 more rows