Determine whether majoritarian election requires candidate registration

Determines whether or not candidates must be officially registered prior to the specified majority election.

Usage

requires_candidate_registration(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = lvls(ballot_date, canton, ballot_type = "election"),
  canton = cantons(ballot_date),
  election_nr = 1L
)

Arguments

ballot_date

FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

"2023-06-18"

lvl

Political level. One of "cantonal" or "federal".

canton

FOKUS-covered canton name. One of

"aargau"

election_nr

Election number. A positive integerish scalar.

Value

A logical scalar.

Details

The absence of a candidate registration requirement usually means that every eligible citizen can be elected, i.e. receive valid votes.

See also

Examples

fokus::requires_candidate_registration(ballot_date = "2019-10-20",
                                       lvl = "federal",
                                       canton = "aargau")
#> [1] FALSE