Shorten column names to a maximum length of 32 characters

This is useful for DTA export since Stata has a built-in variable name length limit of 32 characters (see also here).

Usage

shorten_colnames(x, max_n_char = 32L)

Arguments

x

A tabular data object like a data frame or tibble.

max_n_char

Maximum allowed number of characters. Either NULL to skip the check or otherwise an integerish scalar in which case it is ensured that the maximum resulting variable name length doesn't exceed it. Doesn't have any influence on the applied shortening logic.

Value

x with column names shortened to a maximum length of 32 characters.

See also