Upload files to Google Drive

Uploads one or more files to Google Drive.

Usage

upload_to_g_drive(
  filepaths,
  g_drive_folder,
  path_gcp_service_account_key = Sys.getenv("PATH_GCP_KEY_ZDA"),
  quiet = FALSE
)

Arguments

filepaths

Local path(s) to the file(s) to be uploaded.

g_drive_folder

Destination path on Google Drive where the files are to be uploaded to.

path_gcp_service_account_key

Path to the GCP Service Account Key JSON file. See auth_g_drive_gcp() for details.

quiet

Whether or not to suppress printing status output from googledrive operations.

Value

filepaths, invisibly.

Details

Essentially a convenience wrapper around googledrive::drive_put().

See also