Determine variable's political level(s)

Determines a variable's political level(s).

v_lvls(v_name)

Arguments

v_name
Variable name. A character scalar.

Value

A character vector of political levels, of length 0 if no specific levels could be determined.

Details

Note that the political levels are determined by simply parsing v_name.

See also

Examples

fokus::v_lvls("skill_question_1_cantonal_proposal_1")
#> [1] "cantonal"