Determine variable's proposal number

Determines the proposal number each variable corresponds to. In case no proposal number could be determined for a variable, NA_character is returned.

Usage

var_proposal_nr(var_names)

Arguments

var_names

A character vector of FOKUS variable names.

Value

An integer vector of the same length as var_names.

Details

Note that the determination is performed by simply parsing var_names.

See also