Get referendum proposal combinations for which arguments have been queried

Returns a list containing the political levels, proposal numbers and optionally argument sides (pro/contra) plus argument numbers for all referendum proposal arguments that have been queried in the post-voting survey for the specified ballot date.

Usage

combos_proposal_arguments(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvls = all_lvls,
  canton = cantons(ballot_date),
  incl_arg_side = TRUE,
  incl_arg_nr = incl_arg_side
)

Arguments

ballot_date

FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

"2023-06-18"

lvls

Political level(s). One or more of "cantonal" and "federal".

canton

FOKUS-covered canton name. One of

"aargau"

incl_arg_side

Whether or not to include argument sides (pro/contra) in the resulting list. Setting this to FALSE potentially results in fewer combinations.

incl_arg_nr

Whether or not to include argument numbers in the resulting list. Setting this to FALSE potentially results in fewer combinations. Setting this to TRUE implies incl_arg_side = TRUE.

Value

A list with an element per political-level-proposal-number and optionally -argument-side and -argument-number combination.

See also

Examples

fokus::combos_proposal_arguments(ballot_date = "2023-06-18",
                                 canton = "aargau")
#> [[1]]
#> [[1]]$lvl
#> [1] "cantonal"
#> 
#> [[1]]$proposal_nr
#> [1] 1
#> 
#> [[1]]$argument_side
#> [1] "pro"
#> 
#> [[1]]$argument_nr
#> [1] 1
#> 
#> 
#> [[2]]
#> [[2]]$lvl
#> [1] "cantonal"
#> 
#> [[2]]$proposal_nr
#> [1] 1
#> 
#> [[2]]$argument_side
#> [1] "contra"
#> 
#> [[2]]$argument_nr
#> [1] 1
#> 
#> 
#> [[3]]
#> [[3]]$lvl
#> [1] "cantonal"
#> 
#> [[3]]$proposal_nr
#> [1] 1
#> 
#> [[3]]$argument_side
#> [1] "pro"
#> 
#> [[3]]$argument_nr
#> [1] 2
#> 
#> 
#> [[4]]
#> [[4]]$lvl
#> [1] "cantonal"
#> 
#> [[4]]$proposal_nr
#> [1] 1
#> 
#> [[4]]$argument_side
#> [1] "contra"
#> 
#> [[4]]$argument_nr
#> [1] 2
#> 
#> 
#> [[5]]
#> [[5]]$lvl
#> [1] "cantonal"
#> 
#> [[5]]$proposal_nr
#> [1] 3
#> 
#> [[5]]$argument_side
#> [1] "pro"
#> 
#> [[5]]$argument_nr
#> [1] 1
#> 
#> 
#> [[6]]
#> [[6]]$lvl
#> [1] "cantonal"
#> 
#> [[6]]$proposal_nr
#> [1] 3
#> 
#> [[6]]$argument_side
#> [1] "contra"
#> 
#> [[6]]$argument_nr
#> [1] 1
#> 
#> 
#> [[7]]
#> [[7]]$lvl
#> [1] "cantonal"
#> 
#> [[7]]$proposal_nr
#> [1] 3
#> 
#> [[7]]$argument_side
#> [1] "pro"
#> 
#> [[7]]$argument_nr
#> [1] 2
#> 
#> 
#> [[8]]
#> [[8]]$lvl
#> [1] "cantonal"
#> 
#> [[8]]$proposal_nr
#> [1] 3
#> 
#> [[8]]$argument_side
#> [1] "contra"
#> 
#> [[8]]$argument_nr
#> [1] 2
#> 
#> 
#> [[9]]
#> [[9]]$lvl
#> [1] "cantonal"
#> 
#> [[9]]$proposal_nr
#> [1] 3
#> 
#> [[9]]$argument_side
#> [1] "pro"
#> 
#> [[9]]$argument_nr
#> [1] 3
#> 
#> 
#> [[10]]
#> [[10]]$lvl
#> [1] "federal"
#> 
#> [[10]]$proposal_nr
#> [1] 2
#> 
#> [[10]]$argument_side
#> [1] "pro"
#> 
#> [[10]]$argument_nr
#> [1] 1
#> 
#> 
#> [[11]]
#> [[11]]$lvl
#> [1] "federal"
#> 
#> [[11]]$proposal_nr
#> [1] 2
#> 
#> [[11]]$argument_side
#> [1] "contra"
#> 
#> [[11]]$argument_nr
#> [1] 1
#> 
#> 
#> [[12]]
#> [[12]]$lvl
#> [1] "federal"
#> 
#> [[12]]$proposal_nr
#> [1] 2
#> 
#> [[12]]$argument_side
#> [1] "pro"
#> 
#> [[12]]$argument_nr
#> [1] 2
#> 
#> 
#> [[13]]
#> [[13]]$lvl
#> [1] "federal"
#> 
#> [[13]]$proposal_nr
#> [1] 2
#> 
#> [[13]]$argument_side
#> [1] "contra"
#> 
#> [[13]]$argument_nr
#> [1] 2
#> 
#> 
# without argument numbers
fokus::combos_proposal_arguments(ballot_date = "2023-06-18",
                                 canton = "aargau",
                                 incl_arg_nr = FALSE)
#> [[1]]
#> [[1]]$lvl
#> [1] "cantonal"
#> 
#> [[1]]$proposal_nr
#> [1] 1
#> 
#> [[1]]$argument_side
#> [1] "pro"
#> 
#> 
#> [[2]]
#> [[2]]$lvl
#> [1] "cantonal"
#> 
#> [[2]]$proposal_nr
#> [1] 1
#> 
#> [[2]]$argument_side
#> [1] "contra"
#> 
#> 
#> [[3]]
#> [[3]]$lvl
#> [1] "cantonal"
#> 
#> [[3]]$proposal_nr
#> [1] 3
#> 
#> [[3]]$argument_side
#> [1] "pro"
#> 
#> 
#> [[4]]
#> [[4]]$lvl
#> [1] "cantonal"
#> 
#> [[4]]$proposal_nr
#> [1] 3
#> 
#> [[4]]$argument_side
#> [1] "contra"
#> 
#> 
#> [[5]]
#> [[5]]$lvl
#> [1] "federal"
#> 
#> [[5]]$proposal_nr
#> [1] 2
#> 
#> [[5]]$argument_side
#> [1] "pro"
#> 
#> 
#> [[6]]
#> [[6]]$lvl
#> [1] "federal"
#> 
#> [[6]]$proposal_nr
#> [1] 2
#> 
#> [[6]]$argument_side
#> [1] "contra"
#> 
#> 
# without argument sides and numbers
fokus::combos_proposal_arguments(ballot_date = "2023-06-18",
                                 canton = "aargau",
                                 incl_arg_side = FALSE,
                                 incl_arg_nr = FALSE)
#> [[1]]
#> [[1]]$lvl
#> [1] "cantonal"
#> 
#> [[1]]$proposal_nr
#> [1] 1
#> 
#> 
#> [[2]]
#> [[2]]$lvl
#> [1] "cantonal"
#> 
#> [[2]]$proposal_nr
#> [1] 3
#> 
#> 
#> [[3]]
#> [[3]]$lvl
#> [1] "federal"
#> 
#> [[3]]$proposal_nr
#> [1] 2
#> 
#>