Get proportional election's political parties

Returns the questionnaire code as well as different versions of the name of all parties for the specified proportional election.

Usage

election_parties(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = lvls(ballot_date, canton, ballot_type = "election"),
  canton = cantons(ballot_date),
  election_nr = 1L,
  past = FALSE
)

Arguments

ballot_date

FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

"2023-06-18"

lvl

Political level. One of "cantonal" or "federal".

canton

FOKUS-covered canton name. One of

"aargau"

election_nr

Election number. A positive integerish scalar.

past

Whether to extract the current or the predecessor election's parties.

Value

A tibble.

See also

Examples

fokus::election_parties(ballot_date = "2019-10-20",
                        lvl = "federal",
                        canton = "aargau")
#> # A tibble: 15 × 4
#>     code name.de.qstnr                             name.de.run                        name.de.short
#>    <int> <chr>                                     <chr>                              <chr>        
#>  1     1 SVP oder Junge SVP                        Schweizerische Volkspartei         SVP          
#>  2     2 SP oder JUSO                              SP und Gewerkschaften              SP           
#>  3     3 FDP oder Jungfreisinnige                  FDP.Die Liberalen                  FDP          
#>  4     4 CVP                                       CVP                                CVP          
#>  5     5 Grüne oder Junge Grüne                    Grüne Partei                       Grüne        
#>  6     6 GLP oder Junge GLP                        Grünliberale Partei                GLP          
#>  7     7 BDP oder Junge BDP                        Bürgerlich-Demokratische Partei    BDP          
#>  8     8 EVP oder Junge EVP                        Evangelische Volkspartei           EVP          
#>  9     9 EDU                                       Eidgenössisch-Demokratische Union  EDU          
#> 10    10 PPAG (Piratenpartei)                      Piratenpartei                      PPS          
#> 11    11 LOVB (Lösungs-Orientierte Volks-Bewegung) Lösungs-Orientierte Volks-Bewegung LOVB         
#> 12    12 FW AG (Freie Wähler)                      Freie Wähler Aargau                FW AG        
#> 13    13 LS (Luzi Stamm)                           Luzi Stamm                         LS           
#> 14    14 TEAM65+                                   Die Aargauer Seniorenliste         TEAM65+      
#> 15    15 DU (Die Unabhängigen)                     Die Unabhängigen                   DU           

fokus::election_parties(ballot_date = "2019-10-20",
                        lvl = "federal",
                        canton = "aargau",
                        past = TRUE)
#> # A tibble: 15 × 4
#>     code name.de.qstnr                             name.de.run                        name.de.short
#>    <int> <chr>                                     <chr>                              <chr>        
#>  1     1 SVP oder Junge SVP                        Schweizerische Volkspartei         SVP          
#>  2     2 SP oder JUSO                              SP und Gewerkschaften              SP           
#>  3     3 FDP oder Jungfreisinnige                  FDP.Die Liberalen                  FDP          
#>  4     4 CVP oder Junge CVP                        CVP                                CVP          
#>  5     5 Grüne oder Junge Grüne                    Grüne Partei                       Grüne        
#>  6     7 BDP oder Junge BDP                        Bürgerlich-Demokratische Partei    BDP          
#>  7     6 GLP oder Junge GLP                        Grünliberale Partei                GLP          
#>  8     8 EVP oder Junge EVP                        Evangelische Volkspartei           EVP          
#>  9     9 EDU                                       Eidgenössisch-Demokratische Union  EDU          
#> 10    10 PPAG (Piratenpartei)                      Piratenpartei                      PPS          
#> 11    12 SLB (Sozial-Liberale-Bewegung)            Sozial-Liberale-Bewegung           SLB          
#> 12    13 www.Nichtwähler.ch                        www.Nichtwähler.ch                 NW           
#> 13    11 LOVB (Lösungs-Orientierte Volks-Bewegung) Lösungs-Orientierte Volks-Bewegung LOVB         
#> 14    14 IP (Integrale Politik)                    Integrale Politik Aargau           IP           
#> 15    15 Ecopop                                    Ecopop                             EP