Export QR codes with personalized survey URL

Exports a ZIP file that contains a QR code in SVG and in EPS format for each survey participant storing the personalized survey URL to the private FOKUS directory.

export_qr_codes(
  ballot_date = all_ballot_dates,
  canton = cantons(ballot_date),
  upload_to_g_drive = TRUE,
  g_drive_folder = "fokus/aargau/Umfragen/Dateien für Umfrageinstitut/QR-Codes/",
  quiet = TRUE
)

Arguments

ballot_date
A valid FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

canton
A valid FOKUS canton name (lowercase) covered at ballot_date.

upload_to_g_drive
Whether or not to upload the generated files to the Google Drive folder g_drive_folder.

g_drive_folder
Google Drive folder to deploy the generated files to. Ignored if upload_to_g_drive = FALSE.

quiet
Whether or not to suppress printing status output from googledrive operations.

Value

A tibble containing metadata about the contents of the created ZIP archive, invisibly.

See also