Export questionnaire data

Generates the questionnaire tibble, the Markdown questionnaire and optionally a CSV, an HTML and an XLSX version of it, and writes all of them to the private FOKUS directory and optionally deploys them as a static site (local_deploy_path) and uploads them to a Google Drive folder (g_drive_folder).

Usage

export_qstnr(
  ballot_date = all_ballot_dates,
  canton = cantons(ballot_date),
  verbose = FALSE,
  incl_csv = TRUE,
  incl_html = TRUE,
  incl_xlsx = incl_html,
  deploy = TRUE,
  local_deploy_path = pal::pkg_config_val(key = "qstnr.local_deploy_path", pkg =
    this_pkg),
  upload_to_g_drive = TRUE,
  g_drive_folder = "fokus/aargau/Umfragen/Dateien für Umfrageinstitut/Fragebogen/"
)

Arguments

ballot_date

FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

canton

Canton name. One of

"aargau"

verbose

Whether or not to print detailed progress information during questionnaire generation and Google Drive file upload. Note that questionnaire generation takes considerably more time when this is set to TRUE.

incl_csv

Whether or not to also generate and export a CSV version of the questionnaire.

incl_html

Whether or not to also generate and export an HTML version of the questionnaire.

incl_xlsx

Whether or not to also generate and export an XLSX version of the questionnaire.

deploy

Whether or not to deploy the generated files as a static site to the Git repository specified under local_deploy_path.

local_deploy_path

Local filesystem path to the Git repository of the static site to deploy the generated files files to. Ignored if deploy = FALSE.

upload_to_g_drive

Whether or not to upload the generated files to the Google Drive folder g_drive_folder.

g_drive_folder

Google Drive folder to deploy the generated files to. Ignored if upload_to_g_drive = FALSE.

See also