fokus: Provides an API around the FOKUS Post-voting Surveys

fokus allows to (pre-)process and clean the raw data, analyze and visualize the cleaned data, create the questionnaires and provides other utility functions around the FOKUS post-voting surveys by the Centre for Democracy Studies Aarau (ZDA) at the University of Zurich, Switzerland.

Package options

The following package-specific R options are used to control the package's functionality:

name	description	has fallback if unset

fokus.path_private	path to the working directory (the local instance of the fokus_private repository); defaults to the current working directory	✅

fokus.global_cache_lifespan	default cache lifespan for all functions taking a cache_lifespan argument; defaults to 30 days	✅

See also

Author

Maintainer: Salim Brüggemann salim-b@pm.me (ORCID)

Other contributors:

Centre for Democracy Studies Aarau (ZDA) info@zdaarau.ch [copyright holder]