Questionnaire data

A tibble containing all FOKUS questionnaires as structured data.

Usage

qstnrs

Format

A tibble.

Details

qstnrs was generated based on the following steps:

gen_qstnr_tibble() was run for all valid combinations of canton and ballot_date, some validation checks were performed, and the results were merged into a single tibble.

List columns were expanded to long format.

Columns question_intro_i and question_intro_j were merged into the single column question_intro.

Column question_full was complemented, i.e. made to fall back on question if NA.

Columns question_common and variable_label_common were complemented, i.e. made to fall back on question_full and variable_label respectively if NA.

Markdown formatting was stripped from all character columns.

See also

Examples

fokus::qstnrs
#> # A tibble: 7,706 × 19
#>    ballot_date canton block                variable_name topic who   question_intro question question_full question_common allow_multiple_answers variable_label
#>    <date>      <chr>  <chr>                <chr>         <chr> <chr> <chr>          <chr>    <chr>         <chr>           <lgl>                  <chr>         
#>  1 2018-09-23  aargau 00_cantonal_statist… ballot_date   Abst… alle  ""             NA       NA            NA              FALSE                  date of the b…
#>  2 2018-09-23  aargau 00_cantonal_statist… ballot_type   Urne… alle  ""             NA       NA            NA              FALSE                  type of the b…
#>  3 2018-09-23  aargau 00_cantonal_statist… ballot_type   Urne… alle  ""             NA       NA            NA              FALSE                  type of the b…
#>  4 2018-09-23  aargau 00_cantonal_statist… ballot_type   Urne… alle  ""             NA       NA            NA              FALSE                  type of the b…
#>  5 2018-09-23  aargau 00_cantonal_statist… last_name     Nach… alle  ""             NA       NA            NA              FALSE                  respondent's …
#>  6 2018-09-23  aargau 00_cantonal_statist… first_name    Vorn… alle  ""             NA       NA            NA              FALSE                  respondent's …
#>  7 2018-09-23  aargau 00_cantonal_statist… middle_name   Weit… alle  ""             NA       NA            NA              FALSE                  respondent's …
#>  8 2018-09-23  aargau 00_cantonal_statist… street        Stra… alle  ""             NA       NA            NA              FALSE                  street name o…
#>  9 2018-09-23  aargau 00_cantonal_statist… street_number Haus… alle  ""             NA       NA            NA              FALSE                  street number…
#> 10 2018-09-23  aargau 00_cantonal_statist… zip_code      Post… alle  ""             NA       NA            NA              FALSE                  zip code of r…
#> # ℹ 7,696 more rows
#> # ℹ 7 more variables: variable_label_common <chr>, response_options <chr>, variable_values <int>, value_labels <chr>, value_scale <chr>,
#> #   randomize_response_options <lgl>, is_mandatory <lgl>