Get a proposal's raw argument supplemental questionnaire data

Returns a structured list of argument data from the proposal-specific raw supplemental date-specific FOKUS questionnaire data.

Usage

raw_qstnr_suppl_arguments(
  ballot_date = pal::pkg_config_val("ballot_date"),
  lvl = all_lvls,
  canton = pal::pkg_config_val("canton"),
  proposal_nr = 1L
)

Arguments

ballot_date

FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

"2023-06-18"

"2024-10-20"

lvl

Political level. One of "cantonal" or "federal".

canton

FOKUS-covered canton name. One of

"aargau"

proposal_nr

Proposal number. A positive integerish scalar.

Value

A strict list.

See also

Examples

fokus:::raw_qstnr_suppl_arguments(ballot_date = "2018-11-25",
                                  lvl = "cantonal",
                                  canton = "aargau",
                                  proposal_nr = 2)
#> [[1]]
#> [[1]]$de
#> [[1]]$de$long
#> [1] "Es macht keinen Sinn, dass Auslandschweizer/innen an Nationalratswahlen teilnehmen dürfen, aber nicht an den gleichzeitig stattfindenden Ständeratswahlen."
#> 
#> [[1]]$de$short
#> [1] "Unsinnig, dass Auslandschweizer/innen an Nationalrats-, nicht aber Ständeratswahlen teilnehmen dürfen"
#> 
#> 
#> [[1]]$en
#> [[1]]$en$long
#> [1] "It makes no sense for Swiss citizens living abroad to be allowed to take part in the National Council elections, but not in the Council of States elections that take place at the same time."
#> 
#> 
#> [[1]]$nr
#> [1] 1
#> 
#> [[1]]$side
#> [1] "pro"
#> 
#> 
#> [[2]]
#> [[2]]$de
#> [[2]]$de$long
#> [1] "Wer in den Ständerat gewählt werden kann, soll auch mit den Gegebenheiten des Kantons vertraut sein. Das ist bei jemandem, der im Ausland wohnhaft ist, nicht der Fall."
#> 
#> [[2]]$de$short
#> [1] "Im Ausland wohnhafte sind nicht mit den Gegebenheiten des Kantons vertraut"
#> 
#> 
#> [[2]]$en
#> [[2]]$en$long
#> [1] "Anyone who can be elected to the Council of States should also be familiar with the circumstances of the canton. This is not the case for someone who lives abroad."
#> 
#> 
#> [[2]]$nr
#> [1] 1
#> 
#> [[2]]$side
#> [1] "contra"
#> 
#> 
#> [[3]]
#> [[3]]$de
#> [[3]]$de$long
#> [1] "Die Einführung des Ständeratswahlrechts für Auslandschweizer/innen ist gerechtfertigt, weil nur sehr geringe Mehrkosten entstehen."
#> 
#> [[3]]$de$short
#> [1] "Mehrkosten sind sehr gering"
#> 
#> 
#> [[3]]$en
#> [[3]]$en$long
#> [1] "The introduction of the right to elect members of the Council of States for Swiss citizens living abroad is justified because the additional costs are very low."
#> 
#> 
#> [[3]]$nr
#> [1] 2
#> 
#> [[3]]$side
#> [1] "pro"
#> 
#> 
#> [[4]]
#> [[4]]$de
#> [[4]]$de$long
#> [1] "Über kantonale Angelegenheiten sollen nur jene entscheiden dürfen, die auch von kantonalen Entscheiden betroffen sind. Das ist bei Schweizern und Schweizerinnen, die im Ausland wohnen, nicht der Fall."
#> 
#> [[4]]$de$short
#> [1] "Nur direkt Betroffene sollen über kantonale Angelegenheiten entscheiden"
#> 
#> 
#> [[4]]$en
#> [[4]]$en$long
#> [1] "Only those who are also affected by cantonal decisions should be allowed to decide on cantonal matters. This is not the case for Swiss nationals living abroad."
#> 
#> 
#> [[4]]$nr
#> [1] 2
#> 
#> [[4]]$side
#> [1] "contra"
#> 
#>