Read in easyvote municipality data

Reads in the latest dataset of easyvote municipality information provided to us not earlier than 90 days before and up until 20 days after the ballot_date.

Usage

read_easyvote_municipalities(ballot_date, canton)

Arguments

ballot_date

FOKUS-covered ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

"2023-06-18"

canton

Canton name. One of

"aargau"

Value

A tibble.

Details

If both columns min_age and max_age are NA in the data returned, it means that

the municipality did not provide easyvote with specific information on the target age range, and

the municipality has subscribed to parcel mailing (instead of direct delivery to households) and delivers the brochures itself -- very likely to young adults between 18--25 years.

Examples

# private FOKUS directory needs to be accessible for this function to work
try(
  fokus:::read_easyvote_municipalities(ballot_date = "2020-09-27",
                                       canton = "aargau")
)
#> Error in (function (cond)  : 
#>   error in evaluating the argument 'x' in selecting a method for function 'as_date': Please set the fokus package configuration option path_repo_private by either setting the R option fokus.path_repo_private or the environment variable
#> `FOKUS_PATH_REPO_PRIVATE`.