Determine whether majoritarian election requires candidate registration

Determines whether or not candidates must be officially registered prior to the specified majority election.

requires_candidate_registration(
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  canton = cantons,
  election_nr = 1L
)

Arguments

ballot_date
A valid FOKUS-covered cantonal ballot date. One of

"2018-09-23"

"2018-11-25"

"2019-10-20"

"2020-09-27"

"2020-10-18"

"2021-11-28"

lvl
Political level. One of

"cantonal"

"federal"

canton
A valid FOKUS canton. One of

"aargau"

election_nr
Election number. A positive integer scalar (in almost all cases 1L).

Value

A logical scalar.

Details

The absence of a candidate registration requirement usually means that every eligible citizen can be elected, i.e. receive valid votes.

See also

Examples

fokus::requires_candidate_registration(ballot_date = "2019-10-20",
                                       lvl = "federal",
                                       canton = "aargau")

#> [1] FALSE