Write file to private FOKUS repository

Uploads a file to the private FOKUS repository via GitLab's RESTful API (v4).

Usage

write_private_file(
  path,
  content,
  from_file = FALSE,
  overwrite = TRUE,
  commit_message = "auto: update file via fokus R pkg",
  branch = repo_private_default_branch,
  auth_token = pal::pkg_config_val(key = "token_repo_private", pkg = this_pkg),
  quiet = FALSE
)

Arguments

path

File path relative to the repository root.

content

File content, as a character scalar for text files, or a raw vector for binary files. Or the path to a local file as a character scalar if from_file = TRUE.

from_file

Whether or not content indicates the path to a local file instead of the actual file content.

overwrite

Whether or not to overwrite an already existing file.

commit_message

Git commit message for file creation/update.

branch

Git branch name to upload the file to.

auth_token

Personal access token of a gitlab.com account with access to the private FOKUS repository.

quiet

Whether or not to suppress printing status output from internal processing.

Value

path, invisibly.

See also